32-bit: shift widths, oversized literals, and one real overflow hiding among them
The shift-width class is why `check-portable` exists at all -- it was
built after the `qr` BitWriter bug -- and while the gate was blind the
class came back in five places. Fixed here with the direction chosen per
site rather than by reflex:

  - `iec61850`/`iec62351` BER length encoding and `bls12_381`'s MSM
    window shifts operate on genuine `usize` values, so the shift type
    is derived from the target (`std.math.Log2Int(usize)`) instead of
    hardcoding either width. Hardcoding `u5` would be the same mistake
    mirrored.
  - `rsa`'s bcrypt-KDF `eksBlowfishSetup` goes the other way: `n` is an
    ITERATION COUNT, not a memory size, so it becomes a fixed `u64`.
    Narrowing it to `usize` would have silently capped `cost` at 31 on a
    32-bit host, quietly weakening a KDF -- the worst available outcome.
  - `bls12_381:705`'s `1<<32` assert is an algorithm constant, not a
    fact about the host's address space, so it too becomes `u64`.
  - `dataset`'s `1<<33` test literal is comptime-pruned on 32-bit
    targets; verified it still runs unskipped on 64-bit rather than
    quietly disappearing everywhere.

⚠ One of these was not merely a compile error. `dnp3`'s `item_pos` /
`scan_stride` were a LATENT 32-BIT OVERFLOW: a class-0 scan needs more
than 32 bits of state once the database holds points past the first
`PointKind`. On a 32-bit device that was a wrong answer waiting, and
nothing was going to find it -- the gate that should have was blind, and
no test on a 64-bit host can see it. `fleetsim` clears transitively
through it, with no source change of its own.

Also: `enip`'s hardcoded `1<<63` becomes a target-relative
`maxInt(usize)/2+1`; `kvtree` and `tsdb` widen their splitmix64 mixing
to `u64` to match the field it already fed; `csvstream` and `loopix`
narrow to `usize` matching their own siblings.

Every fix here is compile-only except `dnp3`, which changes behaviour on
32-bit and is stated as such rather than counted as verified.

+++ modules/loopix/src/adversary.zig
@@ -319,7 +319,10 @@ test "measure: a well-mixed transcript (exponential holds, full pool) preserves 
     // Arrivals clustered 0..11; departures deliberately re-ordered vs arrival
     // and spread over ~a few mean-delays so several are plausible for each.
     const deps = [_]Time{ 55, 30, 80, 42, 12, 95, 61, 25, 70, 38, 105, 48 };
-    var i: u64 = 0;
+    // `usize`, not `u64`: this loop indexes the fixed test array `deps[i]`,
+    // and every other use of `i` here (`tr`'s `arr: Time`/`id: u64` params)
+    // widens implicitly from `usize` — there was no actual need for `u64`.
+    var i: usize = 0;
     while (i < deps.len) : (i += 1) {
         try list.append(gpa, tr(0, i, deps[i], i));
     }
@@ -374,7 +377,10 @@ test "reorderStats: order-preserving transcripts score exactly zero" {
     // FifoMix picture. Later arrival can never depart first.
     var list: std.ArrayListUnmanaged(Transit) = .empty;
     defer list.deinit(gpa);
-    var i: u64 = 0;
+    // `usize`, not `u64`: reused below to index `deps_sorted[i]`, and neither
+    // loop needs `u64` range — `tr`'s `arr`/`dep: Time` and `id: u64` params
+    // all widen implicitly from `usize`.
+    var i: usize = 0;
     while (i < 8) : (i += 1) try list.append(gpa, tr(0, i * 10, i * 10 + 40, i));
     const st = reorderStats(list.items);
     try testing.expect(st.pairs >= 10); // plenty of co-resident pairs examined
@@ -400,7 +406,11 @@ test "reorderStats: an interleaved (memoryless-looking) transcript scores near 1
     // The well-mixed synthetic from the `measure` test: clustered arrivals,
     // departures genuinely re-ordered vs arrival.
     const deps = [_]Time{ 55, 30, 80, 42, 12, 95, 61, 25, 70, 38, 105, 48 };
-    var i: u64 = 0;
+    // `usize`, not `u64`: same reasoning as the sibling "well-mixed
+    // transcript" test above — this loop indexes the fixed test array
+    // `deps[i]`, and `tr`'s `arr: Time`/`id: u64` params both widen
+    // implicitly from `usize`, so there was no actual need for `u64`.
+    var i: usize = 0;
     while (i < deps.len) : (i += 1) try list.append(gpa, tr(0, i, deps[i], i));
     const st = reorderStats(list.items);
     try testing.expect(st.pairs >= 30);
